requirements.txt
torch
torchvision
numpy
pandas
pillow
scikit-learn
pyyaml
onnx
onnxruntime
grad-cam
streamlit
pytest
